requirements.txt
appdirs==1.4.3
asn1crypto==0.22.0
awsebcli==3.10.1
blessed==1.14.2
botocore==1.5.43
cement==2.8.2
cffi==1.10.0
colorama==0.3.7
cryptography==1.8.1
Django==1.11
django-cors-headers==2.0.2
django-extensions==1.7.8
django-fernet-fields==0.5
django-filter==1.0.2
djangorestframework==3.6.2
djangorestframework-jwt==1.10.0
docker-py==1.7.2
dockerpty==0.4.1
docopt==0.6.2
docutils==0.13.1
idna==2.5
jmespath==0.9.2
mysqlclient==1.3.10
packaging==16.8
pathspec==0.5.0
pycparser==2.17
PyJWT==1.5.0
pyparsing==2.2.0
python-dateutil==2.6.0
pytz==2017.2
PyYAML==3.12
requests==2.9.1
semantic-version==2.5.0
six==1.10.0
tabulate==0.7.5
termcolor==1.1.0
wcwidth==0.1.7
websocket-client==0.40.0
